Add files via upload
This board game contains two players where one player is the user and the user needs to press enter to roll the dice. The other player is named a computer, the dice are rolled automatically, and a random number between 1-6 rolls for each player. At the end of each game, the program writes a file containing the game's details.